+++ src/App.scss
@@ -2,6 +2,8 @@
     width: 100%;
     min-height: 100vh;
     height: 100%;
+    display: flex;
+    align-items: center;
     // background-color: lightgray;
 }
 
@@ -9,8 +11,10 @@
     width: 100%;
     height: 100%;
     min-height: 100vh;
+    max-width: 800px;
     display: flex;
     justify-content: center;
     align-items: center;
+    overflow: hidden;
     // background-color: salmon;
 }
