--- src/pages/Form.scss
@@ -1,27 +1,0 @@
-.form-container {
-    background-color: lightblue;
-    width: 40%;
-    min-width: 450px;
-    min-height: 400px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-
-
-    form {
-        .form-group {
-            display: flex;
-            flex-direction: column;
-            // justify-content: center;
-            align-items: center;
-
-            // min-height: 300px;
-        }
-
-        .form-controls {
-            vertical-align: bottom;
-        }
-    }
-}
